requirements.txt
aiohttp==3.9.4
aioresponses==0.7.6
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
dnspython==2.4.2
filelock==3.13.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
future==0.18.3
greenlet==3.0.3
httpretty==1.1.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
multidict==6.0.5
pillow==10.3.0
platformdirs==4.2.0
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-Wappalyzer==0.3.1
python-whois==0.9.3
requests==2.31.0
requests-file==1.5.1
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
tldextract==5.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.1.0
virtualenv==20.25.0
Werkzeug==3.0.1
wget==3.2
whois==1.20240129.2
yarl==1.9.4
